requirements.txt
apiai==1.2.3
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
Django==2.0
djangorestframework==3.7.7
idna==2.6
nltk==3.2.5
numpy==1.14.0
pycparser==2.18
pyOpenSSL==17.5.0
pysolr==3.7.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
textblob==0.15.1
urllib3==1.22
watson-developer-cloud==1.0.2
